Converted amount type from double to int (to avoid precision-error while verification of transactions).

+++ src/Account.java
@@ -1,11 +1,11 @@
 public class Account {
 
     private String holderName;
-    private double amount;
+    private int amount;
     private final int id;
     private static int count=0;
 
-    public Account(String holderName, double amount) {
+    public Account(String holderName, int amount) {
         this.holderName = holderName;
         this.amount = amount;
         this.id=++count;
@@ -15,7 +15,7 @@ public class Account {
         return holderName;
     }
 
-    public double getAmount() {
+    public int getAmount() {
         return amount;
     }
 
@@ -27,7 +27,7 @@ public class Account {
         return count;
     }
 
-    public void debit(double delta) throws InsufficientFundsException{
+    public void debit(int delta) throws InsufficientFundsException{
 
         if(delta<amount){
             throw new InsufficientFundsException("Account No. "+id+" has insufficient funds.");
@@ -39,7 +39,7 @@ public class Account {
 
     }
 
-    public void credit(double delta){
+    public void credit(int delta){
         amount=amount+delta;
     }
 }
